+++ ChangeLog.md
@@ -1,1 +1,20 @@
 #Change Log
+
+####2.2.0
+
+#####Added
+######Feature:
+* Add Percent Column In Task Table
+* Title, Time, Progress, Done Columns in Task Table Is Sortable Now, Just Click On Them
+
+<!--
+PATTERN 
+#####Added
+#####Changed
+#####Fixed
+#####Removed
+
+######Feature
+######Improve
+######Bug
+-->